refactor and add a few new elements support
