+++ header.go
@@ -17,7 +17,7 @@ func ExtractOrgHeaders(r *bufio.Reader) (fm []byte, err error) {
 		if err != nil {
 			return nil, err
 		}
-		if p[0] != '#' && p[1] != '+' {
+		if !charMatches(p[0], '#') && !charMatches(p[1], '+') {
 			endOfHeaders = false
 			break
 		}
@@ -31,17 +31,17 @@ func ExtractOrgHeaders(r *bufio.Reader) (fm []byte, err error) {
 	return out.Bytes(), nil
 }
 
+var reHeader = regexp.MustCompile(`^#\+(\w+?): (.*)`)
+
 // OrgHeaders find all of the headers from a byte slice and returns
 // them as a map of string interface
 func OrgHeaders(input []byte) (map[string]interface{}, error) {
-	do := input
 	out := make(map[string]interface{})
-	scanner := bufio.NewScanner(bytes.NewReader(do))
+	scanner := bufio.NewScanner(bytes.NewReader(input))
 
-	reHeader := regexp.MustCompile(`^#\+(\w+?): (.*)`)
 	for scanner.Scan() {
 		data := scanner.Bytes()
-		if data[0] != '#' && data[1] != '+' {
+		if !charMatches(data[0], '#') && !charMatches(data[1], '+') {
 			return out, nil
 		}
 		matches := reHeader.FindSubmatch(data)
