+++ testdata/test.html.golden
@@ -68,29 +68,29 @@
 
 <h6 id="headline-6-4"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span> Headline 6</h6>
 
-<h1 id="headline-1-tags-are-supported">Headline 1  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h1>
+<h1 id="headline-1-tags-are-supported">Headline 1 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h1>
 
-<h2 id="headline-2-tags-are-supported">Headline 2						  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h2>
+<h2 id="headline-2-tags-are-supported">Headline 2 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h2>
 
-<h3 id="headline-3-tags-are-supported">Headline 3 						  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h3>
+<h3 id="headline-3-tags-are-supported">Headline 3 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h3>
 
-<h4 id="headline-4-tags-are-supported">Headline 4						  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h4>
+<h4 id="headline-4-tags-are-supported">Headline 4 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h4>
 
-<h5 id="headline-5-tags-are-supported">Headline 5 					  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h5>
+<h5 id="headline-5-tags-are-supported">Headline 5 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h5>
 
-<h6 id="headline-6-tags-are-supported">Headline 6 					  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h6>
+<h6 id="headline-6-tags-are-supported">Headline 6 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h6>
 
-<h1 id="headline-1-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span> Headline 1  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h1>
+<h1 id="headline-1-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span> Headline 1 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h1>
 
-<h2 id="headline-2-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span> Headline 2 					  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h2>
+<h2 id="headline-2-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span> Headline 2 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h2>
 
-<h3 id="headline-3-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 3 				  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h3>
+<h3 id="headline-3-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 3 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h3>
 
-<h4 id="headline-4-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 4 				  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h4>
+<h4 id="headline-4-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 4 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h4>
 
-<h5 id="headline-5-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 5				  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h5>
+<h5 id="headline-5-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 5 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h5>
 
-<h6 id="headline-6-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 6				  <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h6>
+<h6 id="headline-6-tags-are-supported-1"><span class="todo TODO">TODO</span> <span class="priority A">[A]</span>  Headline 6 <span class="tags">tags</span>  <span class="tags">are</span>  <span class="tags">supported</span> </h6>
 
 <ul><li>unordered </li>
 <li>lists</li>
@@ -107,28 +107,28 @@
 <table>
 <thead>
 <tr>
-<th> tables     </th>
-<th> with     </th>
-<th> headers    </th>
-<th> are      </th>
-<th> supported       </th>
+<th>tables</th>
+<th>with</th>
+<th>headers</th>
+<th>are</th>
+<th>supported</th>
 </tr>
 </thead>
 <tbody>
 <tr>
-<td> cell 1:1   </td>
-<td> cell 1:2 </td>
-<td> cell 1:3   </td>
-<td> cell 1:4 </td>
-<td> cell 1:5        </td>
+<td>cell 1:1</td>
+<td>cell 1:2</td>
+<td>cell 1:3</td>
+<td>cell 1:4</td>
+<td>cell 1:5</td>
 </tr>
 
 <tr>
-<td> <em>emphasis</em> </td>
-<td> <strong>bold</strong>   </td>
-<td> <code>verbatim</code> </td>
-<td> <code>code</code>   </td>
-<td> <del>strikethrough</del> </td>
+<td><em>emphasis</em></td>
+<td><strong>bold</strong></td>
+<td><code>verbatim</code></td>
+<td><code>code</code></td>
+<td><del>strikethrough</del></td>
 </tr>
 </tbody>
 </table>
@@ -136,27 +136,27 @@
 <table>
 <tbody>
 <tr>
-<td> tables     </td>
-<td> without  </td>
-<td> headers    </td>
-<td> are      </td>
-<td> supported       </td>
+<td>tables</td>
+<td>without</td>
+<td>headers</td>
+<td>are</td>
+<td>supported</td>
 </tr>
 
 <tr>
-<td> cell 1:1   </td>
-<td> cell 1:2 </td>
-<td> cell 1:3   </td>
-<td> cell 1:4 </td>
-<td> cell 1:5        </td>
+<td>cell 1:1</td>
+<td>cell 1:2</td>
+<td>cell 1:3</td>
+<td>cell 1:4</td>
+<td>cell 1:5</td>
 </tr>
 
 <tr>
-<td> <em>emphasis</em> </td>
-<td> <strong>bold</strong>   </td>
-<td> <code>verbatim</code> </td>
-<td> <code>code</code>   </td>
-<td> <del>strikethrough</del> </td>
+<td><em>emphasis</em></td>
+<td><strong>bold</strong></td>
+<td><code>verbatim</code></td>
+<td><code>code</code></td>
+<td><del>strikethrough</del></td>
 </tr>
 </tbody>
 </table>
@@ -172,3 +172,7 @@
 <center>
 This is a centered block.
 </center>
+
+<hr />
+
+<p>^ that's support for horizontal rules!</p>
